프로필 설정/회원가입 API 수정
