+++ src/main/java/com/umc/meetpick/controller/ProfileModifyController.java
@@ -1,0 +1,9 @@
+package com.umc.meetpick.controller;
+
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+
+@Tag(name="프로필 수정 관련 API", description = "사용자 프로필 수정 API 입니다.")
+public class ProfileModifyController {
+
+}
