+++ src/main/java/com/umc/meetpick/dto/ProfileModifyDTO.java
@@ -1,0 +1,4 @@
+package com.umc.meetpick.dto;
+
+public class ProfileModifyDTO {
+}
