--- src/main/java/com/umc/meetpick/dto/SignupSuccessDTO.java
@@ -1,11 +1,0 @@
-package com.umc.meetpick.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class SignupSuccessDTO {
-    private Long memberId;
-    private String message;
-}
