+++ src/main/java/com/umc/meetpick/service/ProfileModifyService.java
@@ -1,0 +1,12 @@
+package com.umc.meetpick.service;
+
+import com.umc.meetpick.dto.ProfileModifyDTO;
+import com.umc.meetpick.common.response.ApiResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class ProfileModifyService {
+
+}
